feat(manager-react-components): migrate TagsModal to ODS v19
Migrate TagsModal to ODS v19 and integrate it as part of TagsList component
Use Modal component from MRC instead of ODS modal directly
Unit and Integration test coverage

ref: #MANAGER-18968


BREAKING CHANGE: Tags Modal component is made internal to MRC Library
Mandatory modalHeading prop for TagsList component
onClick prop is removed from TagsList component
