+++ packages/manager-react-components/src/components/index.ts
@@ -40,4 +40,3 @@ export * from './order';
 export * from './badge';
 export * from './modal';
 export * from './tags-list';
-export * from './tags-modal';
