+++ packages/manager-react-components/src/components/tags-list/tags-modal/TagsModal.props.tsx
@@ -1,0 +1,10 @@
+import { ModalOpenChangeDetail } from '@ovhcloud/ods-react';
+
+export type TagsModalProps = {
+  heading: string;
+  open: boolean;
+  tags: string[];
+  onCancel: () => void;
+  onEditTags?: () => void;
+  onOpenChange?: (detail?: ModalOpenChangeDetail) => void;
+};
