+++ packages/manager-react-components/src/components/tags-list/tags-modal/translations/Messages_fr_FR.json
