+++ packages/manager-react-components/src/components/tags-list/tags-modal/translations/index.ts
