+++ packages/manager-react-components/src/components/tags-list/tags-stack/TagsStack.constants.ts
