+++ packages/manager-react-components/src/components/tags-list/tags-stack/TagsStack.props.ts
@@ -1,0 +1,5 @@
+export type TagsStackProps = {
+  tags: string[];
+  maxLines?: number;
+  onClick?: () => void;
+};
