+++ packages/manager-react-components/src/components/tags-list/tags-stack/TagsStack.type.ts
@@ -1,6 +1,4 @@
 import { ElementRef } from 'react';
 import { Badge } from '@ovhcloud/ods-react';
 
-export type TagsObj = { [key: string]: string };
-
 export type HTMLBadgeElement = ElementRef<typeof Badge>;
