--- packages/manager-react-components/src/components/tags-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './tags-modal.component';
