+++ packages/manager-react-components/tsconfig.strict.json
@@ -10,7 +10,8 @@
     "./src/components/Breadcrumb/Breadcrumb.props.ts",
     "./src/components/Breadcrumb/Breadcrumb.component.tsx",
     "./src/hooks/breadcrumb/useBreadcrumb.tsx",
-    "./src/components/tags-list/*",
+    "./src/components/tags-list/**/*",
+    "./src/components/tags-list/tags-modal/translations/*.json",
     "./src/components/modal/*",
     "./src/components/tile/*",
     "./src/components/tile/tile-divider/*",
