--- packages/manager-wiki/stories/manager-react-components/components/TagsModal/documentation.mdx
@@ -1,65 +1,0 @@
-import { Meta } from '@storybook/addon-docs';
-import * as ModalStories from './tags-modal.stories';
-
-<Meta title="Manager React Components/Components/TagsModal/Documentation" />
-
-# Overview
-
-The `TagsModal` component displays a modal window that shows and allows searching through **tags** (key-value pairs) associated with an item (e.g., a customer, resource, etc.).  
-It leverages OVHcloud’s ODS React components for consistent design and behavior.
-
----
-
-## ✨ Props
-
-| Prop                 | Type                                  | Description                                                                 |
-| -------------------- | ------------------------------------- | --------------------------------------------------------------------------- |
-| `displayName`        | `string`                              | The name or label displayed in the modal header.                            |
-| `isOpen`            | `boolean`                             | If `true`, the modal is visible.                                            |
-| `tags`              | `{ [key: string]: string }`           | A dictionary of tags to display (key-value pairs).                          |
-| `displayInternalTags` | `boolean` *(optional)*              | Whether to display internal tags. *(currently not used directly inside)*    |
-| `onCancel`          | `() => void`                          | Callback triggered when the user closes the modal (or clicks **Back**).     |
-| `onEditTags`        | `() => void`                          | Callback triggered when the user clicks the **Edit Tags** button.           |
-
----
-
-## 🔍 Features
-
-✅ Displays a scrollable list of tags  
-✅ Allows searching tags by key or value  
-✅ Includes action buttons to **go back** or **edit tags**
-
----
-
-## 📦 Usage Example
-
-```tsx
-import React, { useState, useRef } from 'react';
-import { TagsModal } from '@ovh-ux/manager-react-components';
-
-export const Example = () => {
-  const modalRef = useRef(null);
-  const [isOpen, setIsOpen] = useState(true);
-
-  const tags = {
-    environment: 'production',
-    owner: 'dev-team',
-    priority: 'high',
-  };
-
-  return (
-    <div>
-      <button onClick={() => setIsOpen(true)}>Open Tags Modal</button>
-      <TagsModal
-        ref={modalRef}
-        isOpen={isOpen}
-        displayName="My Resource"
-        tags={tags}
-        onCancel={() => setIsOpen(false)}
-        onEditTags={() => {
-          console.log('Edit tags clicked');
-        }}
-      />
-    </div>
-  );
-};
